requirements.txt
python-dotenv
pydantic
pydantic-settings
google-genai